better logic
