+++ pymc4/flow/__init__.py
@@ -3,6 +3,7 @@ from .executor import SamplingExecutor, SamplingState
 from .transformed_executor import TransformedSamplingExecutor
 from .posterior_predictive_executor import PosteriorPredictiveSamplingExecutor
 from .meta_executor import MetaSamplingExecutor, MetaPosteriorPredictiveSamplingExecutor
+from .smc_executor import SMCSamplingExecutor
 
 __all__ = [
     "SamplingExecutor",
@@ -10,6 +11,7 @@ __all__ = [
     "PosteriorPredictiveSamplingExecutor",
     "MetaSamplingExecutor",
     "MetaPosteriorPredictiveSamplingExecutor",
+    "SMCSamplingExecutor",
     "evaluate_model",
     "evaluate_model_transformed",
     "evaluate_model_posterior_predictive",
@@ -22,3 +24,4 @@ evaluate_model_transformed = TransformedSamplingExecutor()
 evaluate_model_posterior_predictive = PosteriorPredictiveSamplingExecutor()
 evaluate_meta_model = MetaSamplingExecutor()
 evaluate_meta_posterior_predictive_model = MetaPosteriorPredictiveSamplingExecutor()
+evaluate_model_smc = SMCSamplingExecutor()
